Add files via upload
This was an Exam for OOP at the university, you can add astronomers and stars in .txt files, it will open as many windows as astronomers. You can add stars and all the astronomers will see the modifications made in real-time. You can also visualize all the stars using the "VIEW" button, also you can search for stars in the table with the "SEARCH" button, which has partial matching. All the files are saved and modified in each .txt file provided.